requirements.txt
mcp==1.26.0
google-auth>=2.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
